system-design-challenges-50/day19/pyproject.toml
[tool.poetry]
name = "day19-distributed-file-storage"
version = "1.0.0"
description = "A distributed file storage system with WAL, replication, and recovery"
authors = ["System Designer <designer@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
sqlalchemy = "^1.4.0"
pydantic = "^1.8.0"
python-json-logger = "^2.0.0"
celery = "^5.0.0"
redis = "^3.5.0"
prometheus-client = "^0.11.0"
opentelemetry-api = "^1.3.0"
opentelemetry-sdk = "^1.3.0"
opentelemetry-exporter-otlp = "^1.3.0"
opentelemetry-instrumentation-fastapi = "^0.22b0"
opentelemetry-instrumentation-sqlalchemy = "^0.22b0"
boto3 = "^1.18.0"
google-cloud-storage = "^1.42.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.0"
pytest-mock = "^3.6.0"
aiohttp = "^3.7.0"
requests = "^2.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88